pyproject.toml
[tool.poetry]
name = "diplom"
version = "0.1.0"
description = ""
authors = ["danilu <danilu2537@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
serializers = "^0.2.4"
social-auth-app-django = "^5.2.0"
django-filter = "^23.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
skip-string-normalization = true
skip-magic-trailing-comma = true
include = '\.pyi?$'
exclude = '''
/(
  | \.git
  | \.mypy_cache
  | \.venv
  | migrations
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
line_length = 100
combine_as_imports = true
combine_star = true
skip_glob = '''
  .git,
  .mypy_cache,
  .venv,
  migrations
'''
